Prevent IDOR (Insecure Direct Object Reference) pada Endpoint Berbasis ID

+++ app/Http/Controllers/CMS/ArticleController.php
@@ -78,6 +78,9 @@ class ArticleController extends AppBaseController
             return redirect(route('articles.index'));
         }
 
+        // IDOR Prevention: Authorization check
+        $this->authorize('view', $article);
+
         return view('articles.show')->with('article', $article);
     }
 
@@ -92,6 +95,9 @@ class ArticleController extends AppBaseController
 
             return redirect(route('articles.index'));
         }
+
+        // IDOR Prevention: Authorization check
+        $this->authorize('update', $article);
 
         return view('articles.edit', $this->getOptionItems($id))->with('article', $article);
     }
@@ -108,6 +114,10 @@ class ArticleController extends AppBaseController
 
             return redirect(route('articles.index'));
         }
+
+        // IDOR Prevention: Authorization check
+        $this->authorize('update', $article);
+
         $input = $request->all();
         $removeThumbnail = $request->get('remove_thumbnail');
         if ($request->file('foto')) {
@@ -138,6 +148,9 @@ class ArticleController extends AppBaseController
 
             return redirect(route('articles.index'));
         }
+
+        // IDOR Prevention: Authorization check
+        $this->authorize('delete', $article);
 
         $this->articleRepository->delete($id);
         if (request()->ajax()) {
